Writer cut 7: the measure state, made right (re-homed and 64-bit)
Two questions from Marc, two fixes. First: the RawLengths cache did NOT
traverse writers - it was homed on ProtoWriter, but the codebase's
established home for cross-writer measurement state is NetObjectCache: the
buffer-writer's null-writer sidecar shares the parent's instance by
construction ("share the *same* known objects key"), and MeasureState's
Serialize hands the measuring writer's cache to the writing writer via
netCache.InitializeFrom. RawLengths moved there, so a length measured during
the classic engine's prefix measure now serves the real write all the way
down, and the global Measure(value)->Serialize(output) flow writes entirely
from cache hits - captured the first time. Clearing rides _knownLengths'
lifecycle, making the staleness guarantees identical to what has shipped
for years.

Second: the IMeasuringSerializer reply stays int - the engine treats
non-positive as "measure by traversal" and the traversal is long-capable, so
the generated implementation spills bodies wider than int.MaxValue to it -
but OUR arithmetic was int, which a colossal single body would overflow
silently where classic is correct. Measure_ now accumulates long, the
count-folds multiply in 64-bit, the prefix sites use WriteRawVarint64 (byte
identical for small values), and RawLengths is Dictionary<object, long>,
matching _knownLengths - all changed in place while [PBN9002]-unshipped. A
Measure64 DIM (per-TFM, absent down-level) is parked in the plan doc as the
recorded route if a 64-bit member is ever wanted; the int+spill shape makes
it unnecessary.

Gates: goldens 99/99 twice (55 rewrote); conformance 663/663; differential
exit-green over the rebuilt corpus; SchemaTests 556/556 both TFMs;
protobuf-net.Test 1046/1046 both TFMs; Examples green both TFMs; AotSmoke
Release build clean; DownLevelSmoke PASSED; bench compiles against the new
shapes. The committed Reflection model took the documented one-build-behind
hand-nudge and regenerated clean.

+++ src/NanoBench/DescriptorSerializeBenchmarks.cs
@@ -40,8 +40,8 @@ public class DescriptorSerializeBenchmarks
     private MemoryStream _ms = null!;
     private Meta.TypeModel _protogenModel = null!;
 
-    private delegate int GeneratedMeasure(Model.FileDescriptorSet value, int depth,
-        System.Collections.Generic.Dictionary<object, int> lengths);
+    private delegate long GeneratedMeasure(Model.FileDescriptorSet value, int depth,
+        System.Collections.Generic.Dictionary<object, long> lengths);
     private static readonly GeneratedMeasure s_generatedMeasure = ResolveMeasure();
 
     // reference identity, both TFMs (the BCL's ReferenceEqualityComparer is net5+ only)
@@ -53,7 +53,7 @@ public class DescriptorSerializeBenchmarks
             => System.Runtime.CompilerServices.RuntimeHelpers.GetHashCode(obj);
     }
 
-    private readonly System.Collections.Generic.Dictionary<object, int> _measureScratch = new(RefComparer.Instance);
+    private readonly System.Collections.Generic.Dictionary<object, long> _measureScratch = new(RefComparer.Instance);
 
     [GlobalSetup]
     public void Setup()
@@ -156,7 +156,7 @@ public class DescriptorSerializeBenchmarks
     }
 
     [Benchmark]
-    public int NanoGeneratedMeasure()
+    public long NanoGeneratedMeasure()
     {
         // cleared per invoke: the row prices a cold measure INCLUDING cache population,
         // which is what a root serialize actually pays
